Move note actions to their own file.
Because of circular import problems, this commit also moves the error messages from response.js to errors.js
