+++ lib/web/baseRouter.js
@@ -6,17 +6,19 @@ const response = require('../response')
 
 const baseRouter = module.exports = Router()
 
+const errors = require('../errors')
+
 // get index
 baseRouter.get('/', response.showIndex)
 // get 403 forbidden
 baseRouter.get('/403', function (req, res) {
-  response.errorForbidden(res)
+  errors.errorForbidden(res)
 })
 // get 404 not found
 baseRouter.get('/404', function (req, res) {
-  response.errorNotFound(res)
+  errors.errorNotFound(res)
 })
 // get 500 internal error
 baseRouter.get('/500', function (req, res) {
-  response.errorInternalError(res)
+  errors.errorInternalError(res)
 })
