+++ lib/web/imageRouter/index.js
@@ -5,7 +5,7 @@ const formidable = require('formidable')
 
 const config = require('../../config')
 const logger = require('../../logger')
-const response = require('../../response')
+const errors = require('../../errors')
 
 const imageRouter = module.exports = Router()
 
@@ -22,7 +22,7 @@ imageRouter.post('/uploadimage', function (req, res) {
   form.parse(req, function (err, fields, files) {
     if (err || !files.image || !files.image.path) {
       logger.error(`formidable error: ${err}`)
-      response.errorForbidden(res)
+      errors.errorForbidden(res)
     } else {
       logger.debug(`SERVER received uploadimage: ${JSON.stringify(files.image)}`)
