+++ lib/web/middleware/checkURIValid.js
@@ -1,14 +1,14 @@
 'use strict'
 
 const logger = require('../../logger')
-const response = require('../../response')
+const errors = require('../../errors')
 
 module.exports = function (req, res, next) {
   try {
     decodeURIComponent(req.path)
   } catch (err) {
     logger.error(err)
-    return response.errorBadRequest(res)
+    return errors.errorBadRequest(res)
   }
   next()
 }
