+++ lib/web/middleware/tooBusy.js
@@ -2,14 +2,14 @@
 
 const toobusy = require('toobusy-js')
 
-const response = require('../../response')
+const errors = require('../../errors')
 const config = require('../../config')
 
 toobusy.maxLag(config.tooBusyLag)
 
 module.exports = function (req, res, next) {
   if (toobusy()) {
-    response.errorServiceUnavailable(res)
+    errors.errorServiceUnavailable(res)
   } else {
     next()
   }
