+++ lib/web/userRouter.js
@@ -4,7 +4,7 @@ const archiver = require('archiver')
 const async = require('async')
 const Router = require('express').Router
 
-const response = require('../response')
+const errors = require('../errors')
 const config = require('../config')
 const models = require('../models')
 const logger = require('../logger')
@@ -20,7 +20,7 @@ UserRouter.get('/me', function (req, res) {
         id: req.user.id
       }
     }).then(function (user) {
-      if (!user) { return response.errorNotFound(res) }
+      if (!user) { return errors.errorNotFound(res) }
       var profile = models.User.getProfile(user)
       res.send({
         status: 'ok',
@@ -30,7 +30,7 @@ UserRouter.get('/me', function (req, res) {
       })
     }).catch(function (err) {
       logger.error('read me failed: ' + err)
-      return response.errorInternalError(res)
+      return errors.errorInternalError(res)
     })
   } else {
     res.send({
@@ -48,21 +48,21 @@ UserRouter.get('/me/delete/:token?', function (req, res) {
       }
     }).then(function (user) {
       if (!user) {
-        return response.errorNotFound(res)
+        return errors.errorNotFound(res)
       }
       if (user.deleteToken === req.params.token) {
         user.destroy().then(function () {
           res.redirect(config.serverURL + '/')
         })
       } else {
-        return response.errorForbidden(res)
+        return errors.errorForbidden(res)
       }
     }).catch(function (err) {
       logger.error('delete user failed: ' + err)
-      return response.errorInternalError(res)
+      return errors.errorInternalError(res)
     })
   } else {
-    return response.errorForbidden(res)
+    return errors.errorForbidden(res)
   }
 })
 
@@ -78,7 +78,7 @@ UserRouter.get('/me/export', function (req, res) {
     archive.pipe(res)
     archive.on('error', function (err) {
       logger.error('export user data failed: ' + err)
-      return response.errorInternalError(res)
+      return errors.errorInternalError(res)
     })
     models.User.findOne({
       where: {
@@ -107,7 +107,7 @@ UserRouter.get('/me/export', function (req, res) {
           callback(null, null)
         }, function (err) {
           if (err) {
-            return response.errorInternalError(res)
+            return errors.errorInternalError(res)
           }
 
           archive.finalize()
@@ -115,10 +115,10 @@ UserRouter.get('/me/export', function (req, res) {
       })
     }).catch(function (err) {
       logger.error('export user data failed: ' + err)
-      return response.errorInternalError(res)
+      return errors.errorInternalError(res)
     })
   } else {
-    return response.errorForbidden(res)
+    return errors.errorForbidden(res)
   }
 })
